requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20
pytesseract==0.3.13
Pillow==11.2.1   
pdf2image==1.17.0
PyMuPDF==1.26.1
pydantic==2.11.5
groq==0.27.0
motor==3.7.1
opencv-python==4.11.0.86